chat_app/backend/requirements.txt
flask==3.0.2
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
pytz==2025.2
langchain==0.3.23
langchain-ollama==0.3.2
faiss-cpu==1.10.0
langchain-community==0.3.21
cryptography==44.0.2
pymongo[encryption]